Migrate to using a feature_id integer in feature_cluster
This simplifies the code to only needing to store an integer to
represent a feature_id instead of the full (field, value) natural
key. This reduces the amount of memory needed to keep track of
everything, and also will aid future optimisation by keeping
the type signature of the core algorithm function simple - the
core algorithm is insulated from whatever future fanciness we add
for handling values.