Update objective function to consider both features and documents
This means the objective function can be interpreted as the
average internal degree of this grouping of features and documents
in the case of a bipartite graph.

Examining the results, this seems to result in a more even cluster
size distribution, with the clusters with the fewest features
containing more features on average and vice versa with clusters
with more features. The largest cluster captures fewer features
with this tweak.