Fix issue #32
This fixes issue 32 - when adding sites to a stackexchange corpus,
the natural key is (site_id, Post.Id), and a surrogate integer
doc_id is created for indexing purposes. When re-adding a site
to the corpus, a new doc_id was being assigned for the same natural
key, meaning that the index and the corpus were completely out of
sync.

This fix just checks and reuses the doc_id if the natural key
is already present.

closes #32