Refine the clustering algorithm (#55)
Refines the clustering algorithm to enable adjusting the number of clusters down as well as the currently supported up, and also to enable future adjustments to the clustering procedure that allow for more than just feature-cluster moves. This also drops the sub_iteration between partial batches of the whole feature set, and replaces it with probabilistic acceptance of moves that improve the clustering.

Unlike the old procedure which proceeded in two phases of group testing for possible candidate moves, then individual cluster testing within each group, this approach now takes four phases:

    Test each feature's contribution to its own cluster, and calculate the cluster objective.*
    Test adding each feature to groups of clusters, other than the group it's assigned to on that iteration.
    Test each feature against individual clusters in the best scoring group.
    Accept feature moves with only a certain probability, instead of only moving partial sets of features on each sub_iteration. If there are many possible moves for a feature that improve the objective a move is more likely.

The first phase is the most critical, as it allows evaluation of cluster objective scores before working out feature moves. This enables the functionality to slowly dissolve clusters that contribute little to the overall objective.

I haven't systematically evaluated this but it also seems to lead to improved numerical scores of the overall objective, while only taking slightly more time than the current approach.

Additional improvements:

- Additional logging for algorithm progress
- Early termination based on small number of feature moves or small changes in the objective

Outstanding issues:

    Currently this makes the algorithm non-deterministic, unless only a single process is used for the worker pool.