Sample from features to create new clusters
Change from splitting the cluster with the most features to
create a new cluster, to sampling from all movable features. This
means that new clusters can draw from many different features,
and hopefully more likely to create a new useful cluster.

Also, the termination criterion now focuses on the number of feature
moves only, as that is core to the human assessment of the model,
not the numerical objective.