Improve group test systematicity
For the group testing phase, instead of ignoring the features from
clusters in the same group, dispatch a smaller job checking the
features from that cluster against all the other clusters in that
group. This is mainly added to make it possible to concretely
capture all relationships for the next best clusters scoring.