Further refine the clustering algorithm (#59)
* Improve group test systematicity

For the group testing phase, instead of ignoring the features from
clusters in the same group, dispatch a smaller job checking the
features from that cluster against all the other clusters in that
group. This is mainly added to make it possible to concretely
capture all relationships for the next best clusters scoring.

* Move splitting of too-small clusters to _refine_feature_groups

Library users can freely experiment with granularity separate to
the refine_clusters entry point.

* Bugfix: not updating feature_cluster in _refine_feature_groups

This might lead to some odd behaviour when filling empty clusters.