Simplify indexing process
Instead of incremental merging and lots of temporary files/dbs in
a temp directory, have one temporary file, and write all index
segments to that file. Concurrent access is managed by a Lock,
rather than worrying about SQLite's busy handling which is a bit
weird for this kind of write intensive process.