Refine the number of batches and group test parameters slightly
When searching for more nearest neighbours, break down the
feature space into slightly more groups in order to approximately
spend the same time on the group test phase as the feature test
phase.

Also change the top_k size in the group testing phase to 2,
searching slightly more of the feature space on each iteration -
this should make for better results at the cost of slightly
more compute time.