Refine objective slightly
Update the objective calculation to include a difference from the
worst possible score for this set of features.

The new objective is:

    actual-bipartite-degree - worst-possible-bipartite-degree

The worst possible bipartite degree is if none of the terms
overlap at all - subtracting this means we're measuring against
the real baseline.

Including this additional term helps limit the formation of tiny
clusters and giant clusters.